# backend/src/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WAY-Esports API",
    "version": "1.0.0"
  },
  "servers": [{ "url": "/" }],
  "paths": {
...
    "/api/tournaments": {
      "get": {
        "summary": "Список турниров",
        "parameters": [
          { "name": "game", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TournamentList" } } } } }
      },
      "post": {
        "summary": "Создать турнир",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TournamentCreate" } } } },
        "responses": { "201": { "description": "Создано", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Tournament" } } } } }
      }
    },
    "/api/tournaments/batch": {
      "post": {
        "summary": "Массовое создание турниров",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/TournamentCreate" } } } } },
        "responses": { "201": { "description": "Создано" }, "400": { "description": "Ошибка валидации/дубликаты" } }
      }
    },
    "/api/tournaments/{id}": {
      "get": {
        "summary": "Турнир по ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Tournament" } } } }, "404": { "description": "Not found" } }
      },
      "put": {
        "summary": "Обновить турнир",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TournamentUpdate" } } } },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" }, "403": { "description": "Forbidden" } }
      }
    },
    "/api/tournaments/{id}/register": {
      "post": {
        "summary": "Регистрация команды в турнире",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "teamId": { "type": "string" } }, "required": ["teamId"] } } } },
        "responses": { "200": { "description": "OK" }, "400": { "description": "Bad request" }, "403": { "description": "Subscription required" }, "404": { "description": "Not found" } }
      }
    },
    "/api/matches": {
      "get": {
        "summary": "Список матчей",
        "parameters": [
          { "name": "game", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "tournament", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MatchList" } } } } }
      },
      "post": {
        "summary": "Создать матч",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MatchCreate" } } } },
        "responses": { "201": { "description": "Создано" }, "404": { "description": "Tournament not found" } }
      }
    },
...
    "/api/tasks/bulk-register": {
      "post": {
        "summary": "Очередь: массовая регистрация команд в турнир",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tournamentId": { "type": "string" },
                  "teamIds": { "type": "array", "items": { "type": "string" } }
                },
                "required": ["tournamentId", "teamIds"]
              }
            }
          }
        },
        "responses": { "202": { "description": "Queued" }, "400": { "description": "Bad request" }, "404": { "description": "Not found" } }
      }
    },
...
    "/api/teams": {
      "get": {
        "summary": "Список команд",
        "parameters": [
          { "name": "game", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TeamList" } } } } }
      },
      "post": {
        "summary": "Создать команду",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TeamCreate" } } } },
        "responses": { "201": { "description": "Создано" }, "400": { "description": "Имя/тег заняты" } }
      }
    },
...
    "/api/profile": {
      "get": { "summary": "Профиль пользователя", "responses": { "200": { "description": "OK" }, "401": { "description": "Unauthorized" } } },
      "put": {
        "summary": "Обновить профиль",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProfileUpdate" } } } },
        "responses": { "200": { "description": "OK" }, "400": { "description": "Invalid updates" }, "401": { "description": "Unauthorized" } }
      }
    },
...
    "/api/wallet": {
      "get": { "summary": "Получить кошелёк", "responses": { "200": { "description": "OK" }, "404": { "description": "User not found" } } }
    },
...
  },
  "components": {
    "schemas": {
      "Tournament": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "game": { "type": "string" },
          "status": { "type": "string" },
          "startDate": { "type": "string", "format": "date-time" },
          "registeredTeams": { "type": "array", "items": { "type": "string" } },
          "matches": { "type": "array", "items": { "type": "string" } },
          "maxTeams": { "type": "integer" }
        }
      },
      "TournamentList": {
        "type": "object",
        "properties": { "success": { "type": "boolean" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/Tournament" } } }
      },
      "TournamentCreate": {
        "type": "object",
        "properties": { "name": { "type": "string" }, "game": { "type": "string" }, "startDate": { "type": "string", "format": "date-time" }, "maxTeams": { "type": "integer" } },
        "required": ["name", "game", "startDate", "maxTeams"]
      },
      "TournamentUpdate": { "type": "object" },
      "Match": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "game": { "type": "string" },
          "status": { "type": "string" },
          "team1": { "type": "string" },
          "team2": { "type": "string" },
          "tournament": { "type": "string" },
          "startTime": { "type": "string", "format": "date-time" },
          "score": { "type": "object" }
        }
      },
      "MatchList": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/Match" } } } },
      "MatchCreate": { "type": "object", "properties": { "game": { "type": "string" }, "team1": { "type": "string" }, "team2": { "type": "string" }, "tournament": { "type": "string" }, "startTime": { "type": "string", "format": "date-time" } }, "required": ["game", "team1", "team2", "tournament"] },
...
      "Team": { "type": "object", "properties": { "_id": { "type": "string" }, "name": { "type": "string" }, "tag": { "type": "string" }, "game": { "type": "string" }, "members": { "type": "array", "items": { "type": "string" } } } },
      "TeamList": { "type": "object", "properties": { "success": { "type": "boolean" }, "data": { "type": "array", "items": { "$ref": "#/components/schemas/Team" } } } },
      "TeamCreate": { "type": "object", "properties": { "name": { "type": "string" }, "tag": { "type": "string" }, "game": { "type": "string" } }, "required": ["name", "tag", "game"] },
...
      "ProfileUpdate": { "type": "object", "properties": { "gameProfiles": { "type": "array", "items": { "type": "object" } }, "profileLogo": { "type": "string" } } }
    }
  }
}